+++ src/java/model/Provincias.java
@@ -12,31 +12,30 @@ package model;
 public class Provincias {
 
     //Definicion de variables
-    int COD_PROVINCIA;
+    double COD_PROVINCIA;
     String DSC_CORTA_PROVINCIA;
     String DSC_PROVINCIA;
-    int LOG_ACTIVO;
+    double LOG_ACTIVO;
     
     //Metodo constructor
 
     public Provincias() {
     }
 
-    public Provincias(int COD_PROVINCIA, String DSC_CORTA_PROVINCIA, String DSC_PROVINCIA, int LOG_ACTIVO) {
+    public Provincias(double COD_PROVINCIA, String DSC_CORTA_PROVINCIA, String DSC_PROVINCIA, double LOG_ACTIVO) {
         this.COD_PROVINCIA = COD_PROVINCIA;
         this.DSC_CORTA_PROVINCIA = DSC_CORTA_PROVINCIA;
         this.DSC_PROVINCIA = DSC_PROVINCIA;
         this.LOG_ACTIVO = LOG_ACTIVO;
     }
     
-    
     //Metodos SET y GET
 
-    public int getCOD_PROVINCIA() {
+    public double getCOD_PROVINCIA() {
         return COD_PROVINCIA;
     }
 
-    public void setCOD_PROVINCIA(int COD_PROVINCIA) {
+    public void setCOD_PROVINCIA(double COD_PROVINCIA) {
         this.COD_PROVINCIA = COD_PROVINCIA;
     }
 
@@ -56,11 +55,11 @@ public class Provincias {
         this.DSC_PROVINCIA = DSC_PROVINCIA;
     }
 
-    public int getLOG_ACTIVO() {
+    public double getLOG_ACTIVO() {
         return LOG_ACTIVO;
     }
 
-    public void setLOG_ACTIVO(int LOG_ACTIVO) {
+    public void setLOG_ACTIVO(double LOG_ACTIVO) {
         this.LOG_ACTIVO = LOG_ACTIVO;
     }
